Add support for some common useful expansions like wiki-style text
These expansions can be useful when showing user generated message
text, such as in a blog comment sort of situation.  URLs should be
anchors to make navigation easier, paragraph breaks should be kept
as paragraphs, and text that was indented is most likely some sort
of preformatted text that should be displayed as-is.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e6770374c965bba9ee37dc1991d944e3bcb09267