Work around Safari 3.2.1 OpenID login problems
This turns out to not be a browser bug, but is instead a (mis)feature
in the Jetty servlet container, a popular deployment environment for
Gerrit instances.

Jetty by default places a 4096 byte limit on the entire HTTP request.

The OpenID authentication process uses a relatively large cookie
for the RelyingParty, about 1.5 KB, and long URLs with lots of
data encoded in them.  Combined with the standard headers used by
the Safari and Chrome user agents, this deluge of data overflows
Jetty's request limit when the browser returns with its credentials.

When the request buffer overflows, Jetty returns "413 Full Head"
instead of the completion callback the client code is expecting,
and the entire process stalls out with an empty white screen.

By always using a POST request when we redirect the browser to the
OpenID provider site, the OpenID provider may have a much shorter URL
when it finally redirects the browser back to the Gerrit application.
This will create a smaller Referer header, and makes it much less
likely that the request will overflow the 4096 byte limit in Jetty.

Sites using Jetty should still increase the request buffer size
to a larger setting, to further reduce the risk of large requests
overflowing during the OpenID authentication process.

Bug: GERRIT-43
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 1cf727cc25dc3ee3861b9f5d6c3890c9b2113e9d