Only show the progress meter if we haven't reset the table content
Its possible for the entire HTML construction to complete in the
first time slice, setting the table content to the result before
we swap out.  In doing so the stage is set to 1 but the meter is
not yet initialized.  If we initialize the meter as we defer the
command for later completion we will replace the result content
with the meter, and later cause IndexOutOfBoundsException when we
try to set the row items onto each row position.

Bug: GERRIT-94
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 866e1d837dbf8d8377327a26d0b2a87b07e3449a