Change copyright messages in file headers to AOSP
Although the code is primarily authored by Google, we are supported
under the umbrella of the Android Open Source Project, which keeps
track of all contributor agreements for Gerrit Code Review, among
the other projects it hosts.

Switching the copyright header to a generic one pertaining to the
AOSP simplifies things as additional non-Google authored changes
are accepted into files.  The copyright is muddied as it is, as
the AOSP contributor agreement leaves copyright with the original
author, but grants perpetual use and redistribution rights under
the Apache License.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a5057a257b9980e62dc47c6a4e961f773c02ca20