Switch OpenID relying party to openid4java
dyuproject doesn't properly implement the OpenID 2.0 specification.
In particular it doesn't honor the SIGNON vs. SERVER types for a
discovery result.  A recent change in the Google Accounts provider
caused it to more strictly check the requests coming in, and we
couldn't produce a standards compliant SERVER style request due
to dyuproject not honoring the SIGNON vs. SERVER difference.

Everyone uses the openid4java package, and tests against it, as
it is developed by some of the folks who also write the OpenID
standard document, and is just the most widely deployed package
available.  Using it is a reasonably safe bet that it will work
correctly in the wild.

The downside to openid4java is that we must run a snapshot
development version from the 0.9.5 series, as any prior
"stable" version does not function correctly against the
Google Accounts provider (discovery randomly fails).  It
also more than doubles the size of Gerrit's WAR file, due
to its very large number of dependencies.

Bug: GERRIT-101
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: dd95666c3b4ce4d2081a5e045ada4b23ff9ebb54