Simplify the OpenID login code now that the iframe is gone
We can radically simplify the OpenID login implementation by
using a standard RPC interface for discovery, rather than a
more risky JSONP callback in HTML, and entirely remove any of
the code for a JSONP callback post sign-in.

Apparently we can also drop the Safari check-cookie hack as
Safari allows us to set our cookie when returning from the
OpenID provider (it wouldn't permit that before when we had
Google Accounts inside of an iframe).

Bug: GERRIT-102
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a5eaf3583b26b4dcfc9e6e3722ebe3e8d7084f1e