Shorten our OpenID return_to URL by removing unnecessary data
We can redo the discovery against the confirmed identity that we
receive in the response packet, rather than passing the identity
through a gerrit specific request parameter.  This removes one
copy of the identity string from the input parameters, making the
overall return URL from the OpenID provider much shorter.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 020ac8a24ad8ebf626b593330b9fa993a541e68c