Make sure only valid ObjectIds can be passed into git diff-tree
Just to be really paranoid, validate that the strings we received
from the client and are about to inject into the command line look
like a valid SHA-1 string.  This prevents the use of any special
characters, especially Git operators like "^" or "~", and getting
a different output then we planned to give to the client.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 58af10ef5e46c62ce3bf0b3ab1e2bf712d5ee9c2