Work around focus bugs in WebKit based browsers
WebKit based browsers (Safari, Chrome) can't correctly handle a table
that sinks click events when the table is wrapped inside a FocusPanel.
The event queue gets confused and events are delivered to the table
at the wrong coordinate, making it potentially impossible for the user
to click on the row they intended to.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: c6b1353e8f8f4f7a0558ac3c8ac4a2aa0387dbb8