Select better defaults for min/max access rights when adding ACL entry
Picking the highest negative value for the min and the lowest positive
value for the max tends to yield a safer default configuration for a
new ACL entry in any given category.  This limits the risk if the owner
just smacks the Add button and doesn't pay attention to what they were
doing as they did so.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b128b409edbc59c4bc101672ca3f68363ee3df94