ie6: Correct rendering of commit messages
IE6 won't honor an LF inside of a div formatted with "white-space: pre".
Instead, create <br> and <p> tags ourselves to replace these LFs, such
that the HTML renderer would lay out the lines according to the break
tags.

Unfortunately we have to use <p> rather than <br><br> to insert what
should be a double LF, as IE6 collapses the double <br> tags into one,
trying to outsmart the author of the page.

Unfortunately the <p> margins collapse entirely under IE6, so we have
to use padding on the <p> box to get them back the way we expected to
see, but that messes with Firefox (which keeps the margins correctly)
so we have to make the margin 0px to prevent getting double margins
between paragraphs.

Bug: GERRIT-85
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 4380fcd2ededbe849f4665a89670517454bab54a