Allow GWT *.nocache.js to be cached forever if ?content= is supplied
This trick permits the host page to rewrite its <script> reference
to the *.nocache.js to include the SHA-1 or MD-5 hash of the file
content.  By doing so we can allow the browser or edge proxy to
cache the file indefinitely, as the host page would use another URL
if the nocache file was regenerated on a different compilation.

Since the host page for Gerrit (our primary customer) is smaller
than the corresponding *.nocache.js, and the host page can be sent
with gzip while few browsers support gzip encoding on JavaScript,
we can reduce the load time for each new site visit by fetching the
larger content from local cache.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 850dfb96419290503cb598fed94862cd745e9ad2