Cache *.nocache.js and don't cache the host page
Our host page is significantly smaller than the *.nocache.js produced
by the GWT compiler, and since its HTML it can also be transferred by
way of gzip deflate encoding.

We can improve browser loading performance (slightly) by allowing the
larger *.nocache.js to be cached at the edges, and marking the host
page as never cached.  To ensure the most current code is loaded we
embed a SHA-1 hash of the *.nocache.js content within its URL, thus
allowing the CacheControlFilter to mark it as cached for 1 year, and
ensuring that new versions of code will use a different URL, causing
the browser to load the new *.nocache.js version from the server.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 082d64ade8938701a41a62c2f0ce219a10f8f29d