Create a debug mode only method of logging in to Gerrit
Currently the hosted mode debug session is crashing on my 64 bit Ubuntu
desktop anytime I try to use an OpenID login.  As a work around we now
have a servlet that can authenticate the browser as any user identity
desired, provided that the JRE was started with a system property that
permits its usage, and that we were launched from the hosted mode shell
as the servlet is not mapped in the main web.xml.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 49212fabf57badf5deb4e1e0401847ae699ed407