Improve initial page load by embedding user account data into host page
Like the GerritConfig we now embed the user's account information into
the host page if they have a valid account cookie in the request.  This
avoids 2 RPCs during module initialization for a signed-in user, as we
don't need to connect to the AccountService.

The downside is slightly more CPU load on the server when sending out
the host page, as we need to clone then modify the DOM tree, and then
flatten that down for transmission.  The difference is rather minor,
about 3 ms additional time for the host page request.  This is better
than the 150+ ms required for the 2 RPCs to AccountService, so overall
we should see reduced server load.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 47e29c74c4ccc874bec6fc0abb67b0f50aa488c3