Error out if a user forgets to squash when replacing a patch
This condition happens when a user makes a second commit and tries to
replace it onto an existing change, e.g. if they did:

  $ git commit
  $ repo upload
  ... some time later ...

  $ git commit
  $ repo upload --replace .
     ... specify the prior change ...

Gerrit sees that the change depends upon itself, as the latest patch set
has an ancestor that is a prior patch set of the same change.  This is a
very simple case to detect and error out on during upload, ensuring that
the user has an opportunity to correct their usage.

Bug: GERRIT-137
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a186fc5c53e95f65495c3355b203bce0d6e060d5