Improve wiki formatting for preformatted text and lists
We now format a paragraph that contains at least one line with
manual indentation as a pre-formatted text block.  This fixes
the case presented by dbort in GERRIT-134:

==
key_passwords[None] = None # comment explaining why
for k in sorted(keylist):
 if not k:
   continue
 p = subprocess...
==

The entire block is now preformatted.

We also now format paragraphs that contain lines starting with "- "
or "* " as a list, with each line being a unique item in that list.

Bug: GERRIT-134
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2f61f90ed081f53aa451e90c3788992c43de9509