Automatically expand inline comment editors for larger comments
If a comment starts to grow beyond the default box height of 5 lines,
we expand the editor as soon as the user has stopped typing text for
a short period.  This gives us a chance to expand the box height by
an approximation of the number of vertical lines necessary to hold
the current content in a fixed-width font.

Likewise, if a large amount of text is removed from the editor, we
shrink it back down to a smaller size.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f64eda43575823527e4ccf914dd883bb6fd27455