Cache JGit FileHeader and EditList inside of Ehcache
We now cache the FileHeader and EditList information for a given
file difference within an Ehcache, which is a memory and disk based
cache and claims to be faster than streaming this information out
of our local SQL database.  This uses up a lot less disk space,
as we store only the edit list and not the full patch content.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b5ca5fa0a7a320791b51bbccd994b0e9035306de