Fix "null null" user names during schema upgrade from 9 to 10
A couple of commits back we fixed the OpenID account creator to
correctly set the full_name to NULL if the first name and the
last name weren't supplied by the OpenID provider.  However, a
site may have many accounts with this name already stored in
the database.  Since null is not a very typical given name, nor
a typical family name/surname, we should be reasonably OK to do
this global search-replace on the accounts table.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 44916d6fa5a99f7fad73442b1c9593b34d589295