Honor account.preferred_email_address when checking committer lines
A Gerrit administrator might try to create a service accounts
by inserting a row into the "accounts" table, and a row into
"account_ssh_keys", and then use that account to upload a new
change for code review.  For that to work, we also need to have
the committer email address in the email address list.

To avoid needing to create an account_external_ids record just
for the dummy email address, we now honor preferred_email in
the accounts table, in addition to whatever is listed in the
account_external_ids table.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b6878d3b4fb0cce051a40e05e2e9869c2458581e