Move the SSH key cache into Ehcache
This way the cache can be viewed through "gerrit show-caches", but
can also be configured through gerrit.config, if the administrator
wants to apply different caching rules.

We now cache the parsed version of the PublicKey, rather than text
representation.  This makes authentication slightly quicker, as we
only need to perform the equality test on candidates, and don't do
the decoding steps.

Bug: GERRIT-166
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f26d6f30d29c527f590b6b13cf8abcf2377db391