Also scan by preferred email for --reviewers and --cc during receive
This way dummy accounts created to match mailing lists only need a
record in the accounts table and don't need a useless row in the
account_external_ids table.  It gives more DWIMery to the admin,
at the slightly higher cost of an extra SQL query when the input
doesn't match any account in the database.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e76d9126dad78b1a04c1434dd1e2f775fbb3f0b1