Bind our SSH daemon with SO_REUSEADDR
This permits the daemon to come up even if there are connections stuck
in TIME_WAIT or FIN_WAIT2 status within the kernel, which may happen on
a busy server if the daemon is brought down and back up again quickly.

Its potentially risky that you could run two Gerrit daemons on the
same host on the same port and not realize it, so we'll just have to
trust the administrator to not do this.

Bug: GERRIT-164
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 9410f2cf90dbc356f0db2abc74f69693ffcb1544