Detect cases where system_config has too many rows
Sometimes when setting up a new server a user can find a way to
insert multiple rows into the system_config table.  Instead of
honoring one of them seemingly at ranodm (and ignoring the other
rows that may have critical data in them) we now abort at startup
with a description of the problem.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: ed021858d03ef6399eb0fc28e69216a2640aa9b4