Fix file browser in patch that is taller than the window
If the patch screen is taller than the window the file browser popup
didn't position correctly because the movePointerTo method asked the
entire document window to scroll to the table row position.  That is
more often than not causing the document height to extend even longer,
and then the popup goes far below the bottom of the window.

Instead, we only want to scroll the ScrollPanel if the table happens
to be embedded inside of one.  This leaves the browser window alone,
and causes only the ScrollPanel to advance.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: bee7425dc93ab47fcd0f616b7f7c30363df609a7