Remove pointless null assignment from PatchScreen
I never should have submitted this line.  The content table never
discards the file list, so the effective lifecycle of the file list
is the same as the content table, which is also the same as "this"
(the PatchScreen instance).  Setting the field to null has no
purpose, and only confuses others trying to understand the code.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 43af1f6ad4af9e362bcf75315d30eb657d6c321d