Expose the PopupPanel hide command used by GlobalKey
This way applications can setup a binding on the same key that is used
to open the dialog to close it, making that key a toggle key.  By using
the same command class, the help for the dialog will show the various
different methods of closing the dialog aliased under a single entry.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: c80a47113d887406d9733521a50cf9d5e2c817be