Make n/p jump to last/first line if no more hunks are found
A nice feature of Google Mondrian is that instead of just not going
anywhere when you press 'n' at the last edited line in a diff (or
'p' at the first edited line in a diff), it jumps to and highlights
the very end (or beginning) of the file. This is a little more
satisfying as a user than just not going anywhere, since when it
goes nowhere I question whether somehow I don't have focus with
the keyboard or that it's a bug or something.

Bug: GERRIT-187
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 5db3b0a09b0fcf17767bc5a8b2ec0165e65b60f2