Extract the PushOp logic from PushQueue
There is no reason for this to be wedged inside of the PushQueue,
using static methods.  Its cleaner to make the PushOp actually
be the Runnable we inject into the WorkQueue, and let it manage
the push via its own instance members.  Consequently, it is now
a proper top level class.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a0847d55aae40a7c266c4a8a6fbc9c5743e922c7