Automatically replicate all projects at startup
If replication is enabled, schedule all projects for replication when
the daemon starts.  This way if there is any inconsistency between the
local repository and the remotes, it is fixed during startup.  We can
find such cases when a delayed push never finished, because the server
was shutdown while it was still sleeping in the pending queue.

Bug: GERRIT-110
Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 9f9c96d4b551d9a184cd30be81343c375e266fc4