Delete now unnecessary ImportProjectSubmitTypes
This was created for Gerrit 2.0.9 to support moving per-project
configuration out of the project's ~/.git/config into the db,
where it was more easily edited by the web UI.

Since the per-project configuration was previously a hidden and
undocumented feature, its likely nobody else used it except for
me on review.source.android.com, and anyone who has used it has
already upgraded past 2.0.9.  So this is just dead code now.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 92cf3af8e93658a5af47f2d8873e2871882d271c